Refactor: 更新 TimerManager，使其每帧传出高精度的 deltaTime

+++ Cyan-Stars/Assets/Scripts/Framework/Timer/ITimer.cs
@@ -2,6 +2,6 @@
 {
     public interface ITimer
     {
-        void OnUpdate(float deltaTime);
+        void OnUpdate(double deltaTime);
     }
 }
